Auto-derive PartialEq for enums representing constants
This addition is for the macro for wrapping constant values in enums.
This is usually best practice for enums containing only unit variants
and is important for stratisd to be able to compare the enum variants
representing constants that are returned from a method with a specific
variant that is expected.