Change name to optional type
Make crypt_activate_by_token take an optional name as with the
other activation methods. The name should be optional so that
the checking functionality is exposed in this method. None
corresponds to a NULL pointer in libcryptsetup which does not
activate the device but verifies if it can be done.