Expose safe memory functions
This PR has two major components. The first is the implementation of
safe wrappers for memory handling functions exposed in the libcryptsetup
public API as of 2.3. The second is the switch over to SafeMemHandle in
the keyfile function to limit the usage of safe freeing to one specific
module. This way, we will only have one module to look at for proper memory
scrubbing using libcryptsetup.