Change keyslot API to be more granular
Currently we allow CRYPT_SLOT_ANY to be passed into any
method that takes a keyslot ID. This is not legal in some contexts.
By moving the keyslot ID parameter to each individual method,
this gives control over whether a non-negative number which corresponds
to a specific keyslot ID should be passed in or whether CRYPT_SLOT_ANY
can be passed in through the use of Option::None or Option::Some wrapping a
non-negative integer if the user wants to specify a specific slot.