Require clippy pedantic lint ptr_as_ptr
The reason to add this is that it forces the specification of certain
types in situations where an '_' for a type is ambiguous.

It does this by suggesting a change from the use of the "as"
operator to the cast() function. The cast() function is just an
application of the "as" operator, but when it is used, the rust
compiler with warnings enabled requires that the type of the pointer
being cast be known. This helps avoid certain errors where a cast to
void* may be casting a pointer to an incorrect type.

Note that these changes are partly wrong because I used the clippy --fix
option. Being literal, it will change a c_char to an i8 in the
transformed output. It also will not necessarily fill in the correct
type for the '_', that also has to be done by hand. I haven't filled in
all, which is why the clippy task should fail.

Signed-off-by: mulhern <amulhern@redhat.com>