Allow passing None as FEC and hash device when formatting as VERITY
Make the fec_device and the hash_device fields in the CryptParamsVerity
struct an Option<>, so that it's possible to pass NULL in the C
crypt_params_verity struct.

For fec_device, passing NULL to libcryptsetup indicates that no FEC device
should be created. Because currently this is impossible, an FEC device
must be created when using libcryptsetup-rs.

The hash_device option is not used by libcryptsetup when formatting a
verity partition, so allow passing NULL there as well.